+++ packages/cli-next/.eslintrc.js
@@ -146,7 +146,6 @@ module.exports = {
                 ],
             },
         ],
-        'jsdoc/newline-after-description': 'error',
         'jsdoc/no-types': 'error',
 
         'import/first': 'error',