+++ packages/cli-next/CHANGELOG.md
@@ -11,7 +11,8 @@ commit that affects the code in a way that consumers might care about should inc
 
 ## [Unreleased](https://github.com/Airtable/blocks/compare/@airtable/blocks-cli@2.0.4...HEAD)
 
--   No changes.
+-   Upgrade Typescript version to 5.4.5
+-   Perform routine version upgrade to various packages, including `eslint`
 
 ## [2.0.4](https://github.com/Airtable/blocks/compare/@airtable/blocks-cli@2.0.3...@airtable/blocks-cli@2.0.4)